# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Aurora Hotels — Hotel Management System API",
        "description": "Versioned REST API for authentication, bookings, guests, rooms, invoices, payments, and reports (FR-API-001). All responses use consistent API Resource transformers; all errors share a single {\"error\": {code, message, errors?}} envelope (FR-API-003).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Default API server"
        }
    ],
    "paths": {
...
        "/reports/occupancy": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Occupancy rate for a branch on a given date",
                "operationId": "c9070d80dbdd90f8945119ea8e65c5cf",
                "parameters": [
                    {
                        "name": "branch_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "schema": {
                            "description": "Defaults to today",
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Occupancy report",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "date": {
                                                    "type": "string",
                                                    "format": "date"
                                                },
                                                "total_rooms": {
                                                    "type": "integer"
                                                },
                                                "occupied_rooms": {
                                                    "type": "integer"
                                                },
                                                "occupancy_rate": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 0.72
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/reports/revenue": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Revenue for a branch over a date range, broken down by charge type",
                "operationId": "860094fef4d4af10553e20f1f1429913",
                "parameters": [
                    {
                        "name": "branch_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "start",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Revenue report",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "start": {
                                                    "type": "string",
                                                    "format": "date"
                                                },
                                                "end": {
                                                    "type": "string",
                                                    "format": "date"
                                                },
                                                "total_cents": {
                                                    "type": "integer"
                                                },
                                                "by_charge_type": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "charge_type": {
                                                                "type": "string"
                                                            },
                                                            "amount_cents": {
                                                                "type": "integer"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Sanctum personal access token. Send as `Authorization: Bearer {token}`. Tokens are scoped by ability (FR-API-004) — see the abilities returned at login.",
                "bearerFormat": "Opaque token issued by POST /auth/login",
                "scheme": "bearer"
            }
        }
    },
...
}